requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.3
openpyxl==3.0.9
pymongo==3.12.1
python-dotenv==0.19.1
requests==2.26.0
selenium==4.0.0
tqdm==4.62.3
